Limited the imports pof office-ui-fabric to the minimum required for the included templates (#216)
Added more comments to explain the imports
Moved the import of SearchResultsTemplateThemeColors.scss, to bring it in line with how it behaved when using SPFx 1.6.0
Changed SearchResultsTemplateThemeColors.scss to use the theme tokens instead of mixins and variables. For clarity as this does not have an impact on size or behavior.